+++ console/ui/src/main.ts
@@ -1,3 +1,5 @@
+/// <reference types="@angular/localize" />
+
 // Copyright 2020 The Nakama Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");