+++ server/console_notification.go
@@ -23,6 +23,7 @@ import (
 	"slices"
 	"time"
 
+	"github.com/doublemo/nakama-common/api"
 	"github.com/doublemo/nakama-plus/v3/console"
 	"github.com/gofrs/uuid/v5"
 	"github.com/jackc/pgx/v5/pgtype"
@@ -270,6 +271,78 @@ func (s *ConsoleServer) DeleteNotification(ctx context.Context, in *console.Dele
 	if _, err := s.db.ExecContext(ctx, "DELETE FROM notification WHERE id = $1", in.Id); err != nil {
 		s.logger.Error("Error deleting notification.", zap.Error(err))
 		return nil, status.Error(codes.Internal, "failed to delete notification")
+	}
+
+	return &emptypb.Empty{}, nil
+}
+
+func (s *ConsoleServer) SendNotification(ctx context.Context, in *console.SendNotificationRequest) (*emptypb.Empty, error) {
+	if l := len(in.UserIds); l == 0 {
+		senderId := uuid.Nil.String()
+		if in.SenderId != "" {
+			if _, err := uuid.FromString(in.SenderId); err != nil {
+				return nil, status.Error(codes.Internal, "failed to send notification, invalid sender id")
+			}
+			senderId = in.SenderId
+		}
+		contentBytes, err := in.Content.MarshalJSON()
+		if err != nil {
+			s.logger.Error("Error marshaling notification content.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "failed to send notification, invalid content")
+		}
+		notification := &api.Notification{
+			Id:         uuid.Must(uuid.NewV4()).String(),
+			Subject:    in.Subject,
+			Content:    string(contentBytes),
+			Code:       in.Code,
+			SenderId:   senderId,
+			CreateTime: &timestamppb.Timestamp{Seconds: time.Now().UTC().Unix()},
+			Persistent: in.Persistent,
+		}
+
+		if err := NotificationSendAll(ctx, s.logger, s.db, s.tracker, s.router, notification); err != nil {
+			s.logger.Error("Error sending notification.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "failed to send notification")
+		}
+	} else {
+		senderId := uuid.Nil.String()
+		if in.SenderId != "" {
+			if _, err := uuid.FromString(in.SenderId); err != nil {
+				return nil, status.Error(codes.Internal, "failed to send notification, invalid sender id")
+			}
+			senderId = in.SenderId
+		}
+		contentBytes, err := in.Content.MarshalJSON()
+		if err != nil {
+			s.logger.Error("Error marshaling notification content.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "failed to send notification, invalid content")
+		}
+		t := &timestamppb.Timestamp{Seconds: time.Now().UTC().Unix()}
+		notifications := make(map[uuid.UUID][]*api.Notification, l)
+		for _, id := range in.UserIds {
+			userID, err := uuid.FromString(id)
+			if err != nil {
+				s.logger.Error("Error parsing user id.", zap.Error(err), zap.String("id", id))
+				return nil, status.Error(codes.Internal, "failed to send notification, invalid user id")
+			}
+			if userID == uuid.Nil {
+				return nil, status.Error(codes.Internal, "failed to send notification, cannot send to system user")
+			}
+			notifications[userID] = []*api.Notification{{
+				Id:         uuid.Must(uuid.NewV4()).String(),
+				Subject:    in.Subject,
+				Content:    string(contentBytes),
+				Code:       in.Code,
+				SenderId:   senderId,
+				CreateTime: t,
+				Persistent: in.Persistent,
+			}}
+		}
+
+		if err := NotificationSend(ctx, s.logger, s.db, s.tracker, s.router, notifications); err != nil {
+			s.logger.Error("Error sending notification.", zap.Error(err))
+			return nil, status.Error(codes.Internal, "failed to send notification")
+		}
 	}
 
 	return &emptypb.Empty{}, nil