+++ server/console_setting.go
@@ -16,13 +16,13 @@ package server
 
 import (
 	"context"
+	"database/sql"
 	"errors"
 	"slices"
 
 	"go.uber.org/zap"
 
 	"github.com/doublemo/nakama-plus/v3/console"
-	pgx "github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgtype"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -49,7 +49,7 @@ func (s *ConsoleServer) GetSetting(ctx context.Context, in *console.SettingReque
 			return nil, status.Error(codes.Canceled, "Request was canceled.")
 		}
 
-		if errors.Is(err, pgx.ErrNoRows) {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, ErrSettingNotFound
 		}
 
@@ -79,7 +79,7 @@ func (s *ConsoleServer) UpdateSetting(ctx context.Context, in *console.UpdateSet
 		if errors.Is(err, context.Canceled) {
 			return nil, status.Error(codes.Canceled, "Request was canceled.")
 		}
-		if errors.Is(err, pgx.ErrNoRows) {
+		if errors.Is(err, sql.ErrNoRows) {
 			return nil, ErrSettingNotFound
 		}
 		return nil, status.Error(codes.Internal, "An error occurred while trying to update the setting.")